NewGUI: Add contents to Descent profile page.
Localization strings added for descent page.
DescentProfile.cs also modified to be compatible with the NewGUI.

+++ Plugin/TrajectoriesVesselSettings.cs
@@ -5,25 +5,25 @@ namespace Trajectories
     class TrajectoriesVesselSettings : PartModule
     {
         [KSPField(isPersistant = true, guiActive = false)]
-        public float EntryAngle;
+        public double EntryAngle;
 
         [KSPField(isPersistant = true, guiActive = false)]
         public bool EntryHorizon;
 
         [KSPField(isPersistant = true, guiActive = false)]
-        public float HighAngle;
+        public double HighAngle;
 
         [KSPField(isPersistant = true, guiActive = false)]
         public bool HighHorizon;
 
         [KSPField(isPersistant = true, guiActive = false)]
-        public float LowAngle;
+        public double LowAngle;
 
         [KSPField(isPersistant = true, guiActive = false)]
         public bool LowHorizon;
 
         [KSPField(isPersistant = true, guiActive = false)]
-        public float GroundAngle;
+        public double GroundAngle;
 
         [KSPField(isPersistant = true, guiActive = false)]
         public bool GroundHorizon;
